+++ ebench/CMDS.py
@@ -1,2 +1,3 @@
 CMD_RIGOL="ebRigol"
 CMD_UNIT="ebUnit"
+CMD_MENU="ebMenu"