+++ ebench/ebMenu.py
@@ -4,14 +4,15 @@ from .ebench import MenuCtrl
 # Menuactions
 from .ebench import usage, usageCommand, menuStartRecording, menuStopRecording
 
-from . import hello
 
 import yaml
+import sys
 import os
 from absl import app, flags, logging
 from absl.flags import FLAGS
 
 flags.DEFINE_string('config', None, "Path configuration yaml file")
+flags.DEFINE_string('syspath', None, "Append to syspath")
 
 
 # ------------------------------------------------------------------
@@ -46,40 +47,9 @@ defaults = {
     MenuCtrl.MENU_REC_SAVE: { "fileName": "apu.sh"},
 }
 
-# ------------------------------------------------------------------
-# with open( "demo.yaml", "r") as y:
-#    subMenuDefs = yaml.safe_load(y)
-       
-       
-       
-#subMenuDefs = [
-    # {
-    #     MenuCtrl.SUB_MENU_MODULE: "ebench.hello",
-    #     MenuCtrl.SUB_MENU_NAME: MENU_HELLO_MENU,
-    #     MenuCtrl.SUB_MENU_PARAMS: { "initCount": "62"},
-        
-    # },
-    # {
-    #     MenuCtrl.SUB_MENU_MODULE: "ebench.ebRigol",
-    #     MenuCtrl.SUB_MENU_NAME: "Rigol",
-    #     MenuCtrl.SUB_MENU_PARAMS: { "ip": "skooppi"},
-    # },
-    # {
-    #     MenuCtrl.SUB_MENU_MODULE: "ebench.ebUnit",
-    #     MenuCtrl.SUB_MENU_NAME: "UTG962",
-    #     MenuCtrl.SUB_MENU_PARAMS: { "addr": "USB0::0x6656::0x0834::1485061822::INSTR"},
-    # },
-    
-# ]
-
-
-
-# ------------------------------------------------------------------
-# Menu actions
-def hello1( whom:str ):
-    print( "Hello {}!".format(whom))
-
-
+def printExampleYaml():
+    with open(os.path.join( os.path.dirname(__file__), "ebMenu.yaml"), "r") as f:
+        print( f.read())
 
 
 # ------------------------------------------------------------------
@@ -102,26 +72,22 @@ def run( _argv, parentMenu:MenuCtrl=None, config=None):
     cmdController = MenuCtrl( args=_argv, prompt="[?=help, q=quit]" )
 
     # Load submenus con ebMenu.yaml?
-
     subMenuDefs = loadsubMenuDefs(config=config)
 
     subMenus = cmdController.registerSubMenus(subMenuDefs=subMenuDefs)
 
     mainMenu = subMenus | {
-        MENU_HELLO1              : ( "Start hello1", helloPar, hello1),
         # MENU_HELLO_MENU          : ( "Start sub menu", None, subMenu( command=MENU_HELLO_MENU, parentMenu=cmdController, run=hello.run ) ),
         MenuCtrl.MENU_QUIT       : ( "Exit", None, None),
-        "Measurements"           : ( None, None, None ),
-        "measure"                : ( "Measure", MenuCtrl.MENU_INSTRUMENT_ACCESS_PARAMS, cmdController.intrumentAccessMenuAction()),
         "Other"                  : ( None, None, None ),
         MenuCtrl.MENU_HELP       : ( "List commands", None,
-                                    lambda **argV: usage(cmd=CMD, mainMenu=mainMenu, synopsis="Tool to control Rigol MSO1104Z osciloscope")),
+                                    lambda **argV: usage(cmd=CMD, mainMenu=mainMenu, synopsis="Menu controller")),
         MenuCtrl.MENU_HELP_CMD   : ( "List command parameters", MenuCtrl.MENU_HELP_CMD_PARAM,
                                  lambda **argV: usageCommand(mainMenu=mainMenu, **argV )),
 
         MenuCtrl.MENU_REC_START  : ( "Start recording", None, menuStartRecording(cmdController) ),
-        MenuCtrl.MENU_REC_SAVE   : ( "Stop recording", stopRecordingPar, menuStopRecording(cmdController, pgm=_argv[0], fileDir=FLAGS.recordingDir) ),        
-
+        MenuCtrl.MENU_REC_SAVE   : ( "Stop recording", stopRecordingPar, menuStopRecording(cmdController, pgm=_argv[0], fileDir=FLAGS.recordingDir) ),
+        "_yaml"                   : ( "Exaxample yaml", None, printExampleYaml),
         
     }
 
@@ -140,6 +106,9 @@ def run( _argv, parentMenu:MenuCtrl=None, config=None):
 
 def _main( _argv ):
     logging.set_verbosity(FLAGS.debug)
+    if FLAGS.syspath:
+        sys.path.insert( 0, FLAGS.syspath )
+    logging.info( "sys.path={}".format(sys.path))
     run( _argv, config=FLAGS.config)